refactor: 패키지 구조 변경 및 공용 컴포넌트 분리
- 기존 `ui` 패키지 하위에 있던 화면 및 테마 관련 파일들을 기능별 패키지 (`features`) 및 공용 패키지 (`common`)로 재구성했습니다.
- `MyRoutineApp.kt`, `Color.kt`, `Theme.kt`, `Type.kt` 파일을 `common` 패키지 하위로 이동하고, `BottomNavigationBar.kt`를 `common/components` 패키지로 이동했습니다.
- `TodayScreen.kt`는 `features/today` 패키지로 이동했습니다.
- `MainActivity.kt`에서 변경된 `MainScreen`의 경로를 반영했습니다.
