+++ app/src/main/java/com/example/myroutine/MainActivity.kt
@@ -4,6 +4,7 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
+import com.example.myroutine.common.MainScreen
 import com.example.myroutine.ui.theme.MyRoutineTheme
 
 class MainActivity : ComponentActivity() {
