+++ app/src/main/java/com/example/myroutine/common/components/BottomNavigationBar.kt
@@ -14,7 +14,7 @@ import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
 import androidx.navigation.compose.currentBackStackEntryAsState
 import com.example.myroutine.R
-import com.example.myroutine.Routes
+import com.example.myroutine.common.Routes
 
 @Composable
 fun BottomNavigationBar(navController: NavController) {
