+++ app/src/main/java/com/example/myroutine/features/today/TodayScreen.kt
@@ -1,4 +1,4 @@
-package com.example.myroutine.ui.screens
+package com.example.myroutine.features.today
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Row
